+++ controllers/comment.controller.ts
@@ -1,0 +1,119 @@
+import { Request, Response } from 'express';
+
+import httpCode from '../utils/httpcodes';
+import Comment from '../interfaces/Comment.interface';
+import CommentModel from '../models/Comment.model';
+
+class CommentController {
+	/**
+	 *
+	 * @param req express.Request
+	 * @param res express.Response
+	 * @description Create a new post
+	 */
+	static create = async (req: Request, res: Response) => {
+		try {
+			const body: Comment = { ...req.body };
+
+			CommentModel.create(body)
+				.then((result: Comment) => {
+					return res.status(httpCode.CREATED).json(result);
+				})
+				.catch((error) => {
+					return res.status(httpCode.BAD_REQUEST).json(error);
+				});
+		} catch (error) {
+			return res.status(httpCode.INTERNAL_SERVER_ERROR).json(error);
+		}
+	};
+
+	/**
+	 *
+	 * @param req express.Request
+	 * @param res express.Response
+	 * @description Remove a post
+	 */
+	static remove = async (req: Request, res: Response) => {
+		try {
+			const body: { id: string } = req.body;
+
+			CommentModel.deleteOne({ id: body.id })
+				.then((result) => {
+					return res.status(httpCode.ACCEPTED).json(result);
+				})
+				.catch((error) => {
+					return res.status(httpCode.BAD_REQUEST).json(error);
+				});
+		} catch (error) {
+			return res.status(httpCode.INTERNAL_SERVER_ERROR).json(error);
+		}
+	};
+
+	/**
+	 *
+	 * @param req express.Request
+	 * @param res express.Response
+	 * @description Get all posts
+	 */
+	static getFeed = async (req: Request, res: Response) => {
+		try {
+			const body: Comment = { ...req.body };
+
+			CommentModel.find({ ...body })
+				.then((result) => {
+					return res.status(httpCode.ACCEPTED).json(result);
+				})
+				.catch((error) => {
+					return res.status(httpCode.BAD_REQUEST).json(error);
+				});
+		} catch (error) {
+			return res.status(httpCode.INTERNAL_SERVER_ERROR).json(error);
+		}
+	};
+
+	/**
+	 *
+	 * @param req express.Request
+	 * @param res express.Response
+	 * @description Get a post
+	 */
+	static getOne = async (req: Request, res: Response) => {
+		try {
+			const body: { id: string } = req.body;
+
+			CommentModel.findById(body.id)
+				.then((result) => {
+					return res.status(httpCode.ACCEPTED).json(result);
+				})
+				.catch((error) => {
+					return res.status(httpCode.BAD_REQUEST).json(error);
+				});
+		} catch (error) {
+			return res.status(httpCode.INTERNAL_SERVER_ERROR).json(error);
+		}
+	};
+
+	/**
+	 *
+	 * @param req express.Request
+	 * @param res express.Response
+	 * @description Update a post
+	 */
+	static update = async (req: Request, res: Response) => {
+		try {
+			const body: { id: string; post: Comment } = { ...req.body };
+
+			CommentModel.updateOne({ id: body.id }, { ...body.post })
+				.then((result) => {
+					return res.status(httpCode.ACCEPTED).json(result);
+				})
+				.catch((error) => {
+					return res.status(httpCode.BAD_REQUEST).json(error);
+				});
+		} catch (error) {
+			return res.status(httpCode.INTERNAL_SERVER_ERROR).json(error);
+		}
+	};
+}
+
+export default CommentController;