+++ interfaces/Comment.interface.ts
@@ -1,0 +1,12 @@
+interface Comment {
+	id: string;
+	userId: string;
+	parentId: string;
+	type: 'answer' | 'comment' | 'reply';
+	description: string;
+	upvoteCount?: number;
+	downvoteCount?: number;
+	timestamp: number;
+}
+
+export default Comment;