+++ interfaces/Expert.interface.ts
@@ -1,0 +1,12 @@
+interface Expert {
+	id: string;
+	username: string;
+	personalName: string;
+	mail: string;
+	expertizeTech: string;
+	interestedTech: string;
+	description: string;
+	languages: string;
+}
+
+export default Expert;