+++ routes/experts.route.ts
@@ -1,3 +1,6 @@
-import router from "../utils/router";
+import router from '../utils/router';
+import expert from '../controllers/expert.controller';
+import authorize from '../middlewares/authorize.middleware';
 
-router.get("/", (res, req) => {});
+router.get('/expert', expert.getFeed);
+router.get('/expert/:id', authorize, expert.getOne);