+++ routes/post.route.ts
@@ -1,12 +1,11 @@
-import { Request, Response } from 'express';
-
 import router from '../utils/router';
 import authorize from '../middlewares/authorize.middleware';
+import post from '../controllers/post.controller';
 
-router.get('/', (_req: Request, _res: Response) => {});
-
-router.post('/create', authorize, (req: Request, res: Response) => {
-	res.json({ req: req.body });
-});
+router.post('/create', authorize, post.create);
+router.delete('/remove', authorize, post.remove);
+router.get('/get', post.getFeed);
+router.get('/get:id', post.getOne);
+router.put('/update', authorize, post.update);
 
 export default router;