+++ routes/qa.route.ts
@@ -1,3 +1,0 @@
-import router from "../utils/router";
-
-router.get("/", (res, req) => {});