+++ routes/vote.route.ts
@@ -1,0 +1,5 @@
+import router from '../utils/router';
+import authorize from '../middlewares/authorize.middleware';
+import vote from '../controllers/vote.controller';
+
+router.post('/add', authorize, vote.addVote);